full explanation on how to create a dht network
